feat(lunches): add lunches form in point of interest form
- refactored nested js to own files
- fixed nested js for price information forms as there were a typo
- added nested js for lunches
- added forms for lunches with nested forms for dates and lunch offers
- added handling of lunches attributes to point of interest controller
  methods
- updated eslint config to know more globals

SVA-85
