+++ app/javascript/packs/application.js
@@ -14,6 +14,11 @@ require('datatables.net-bs4/js/dataTables.bootstrap4.js');
 require('channels');
 require('@kanety/jquery-nested-form');
 
+require('../partials/_nested_forms');
+// special multi nested forms:
+require('../partials/_nested_content_block_form');
+require('../partials/_nested_lunch_form');
+
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
 // or the `imagePath` JavaScript helper below.
@@ -47,101 +52,6 @@ const initClassicEditor = (htmlEditor) => {
 /* eslint-disable func-names */
 $(function() {
   document.querySelectorAll('.html-editor').forEach((htmlEditor) => initClassicEditor(htmlEditor));
-
-  const defaultNestedFormsOptions = {
-    remover: '.remove',
-    postfixes: '',
-    afterRemoveForm: ($form) => {
-      $form.remove();
-    }
-  };
-
-  $('#nested-categories').nestedForm({
-    forms: '.nested-category-form',
-    adder: '#nested-add-category',
-    ...defaultNestedFormsOptions
-  });
-
-  $('#nested-event-dates').nestedForm({
-    forms: '.nested-event-date-form',
-    adder: '#nested-add-event-dates',
-    ...defaultNestedFormsOptions
-  });
-
-  $('#nested-event-contacts').nestedForm({
-    forms: '.nested-event-contact-form',
-    adder: '#nested-add-event-contacts',
-    ...defaultNestedFormsOptions
-  });
-
-  $('#price_informations').nestedForm({
-    forms: '.price_informations-form',
-    adder: '#nested-add-price_informations',
-    ...defaultNestedFormsOptions
-  });
-
-  $('#nested-web-urls').nestedForm({
-    forms: '.nested-web-url-form',
-    adder: '#nested-add-web-urls',
-    ...defaultNestedFormsOptions,
-    associations: 'urls' // needed to correctly increment ids of added sections
-  });
-
-  // media not nested in a content block, for example in events.
-  // everything with classes here, because in content blocks nested-media will appear multiple times
-  $('.nested-media').nestedForm({
-    forms: '.nested-medium-form',
-    adder: '.nested-add-medium',
-    ...defaultNestedFormsOptions
-  });
-
-  if ($('#nested-content-blocks').length) {
-    const initNestedMediaContents = ($form) => {
-      const timestamp = Date.now();
-
-      $form.find('.nested-media-content-block').addClass(`nested-media-content-block-${timestamp}`);
-      $form.find('.nested-medium-form').addClass(`nested-medium-form-${timestamp}`);
-      $form.find('.nested-add-medium').addClass(`nested-add-medium-${timestamp}`);
-
-      $form.find(`.nested-media-content-block-${timestamp}`).nestedForm({
-        forms: `.nested-medium-form-${timestamp}`,
-        adder: `.nested-add-medium-${timestamp}`,
-        ...defaultNestedFormsOptions,
-        associations: 'media_contents' // needed to correctly increment ids of added sections
-      });
-    };
-
-    $('#nested-content-blocks').nestedForm({
-      forms: '.nested-content-block-form',
-      adder: '#nested-add-content-block',
-      remover: '.removeContent',
-      postfixes: '',
-      afterInitialize: function() {
-        const $initialForms = $('.nested-content-block-form');
-
-        $initialForms.each((index, form) => {
-          initNestedMediaContents($(form));
-        });
-      },
-      beforeAddForm: function($container, $form) {
-        // we only want one initialized media content, so remove eventually created others
-        $form.find('.nested-medium-form').each((index, form) => {
-          if (index > 0) {
-            $(form).remove();
-          }
-        });
-      },
-      afterAddForm: function($container, $form) {
-        initNestedMediaContents($form);
-
-        // init html editors for content block fields body and intro
-        $form
-          .get('0')
-          .querySelectorAll('.html-editor')
-          .forEach((htmlEditor) => initClassicEditor(htmlEditor));
-      }
-    });
-  }
 
   // Init DataTables for all tables with css-class 'data_table'
   $.fn.dataTable = DataTable;
@@ -212,3 +122,4 @@ $(function() {
     e.preventDefault();
   });
 });
+/* eslint-enable func-names */
