+++ app/views/point_of_interests/_form.html.erb
@@ -93,6 +93,8 @@
 
   <%= render partial: "shared/partials/media_contents_form", locals: { record: point_of_interest, form: f } %>
 
+  <%= render partial: "point_of_interests/form_partials/lunches_form", locals: { record: point_of_interest, form: f } %>
+
   <hr />
 
   <div class="row">
